\name{BackTestWeights}
\alias{BackTestWeights}
\title{BackTesting Portfolio Weights Function}
\usage{
  BackTestWeights(returns, rebalance.periods,
    training.length, FUN, ...)
}
\arguments{
  \item{returns}{an xts object of asset returns}
}
\value{
  returns an xts object of portfolio weights
}
\description{
  Calculates rebalance dates
}
\author{
  Doug Martin
}

