% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.Boxplot.R
\name{chart.Boxplot2}
\alias{chart.Boxplot2}
\title{box whiskers plot wrapper}
\usage{
chart.Boxplot2(R, horizontal = TRUE, names = TRUE, as.Tufte = FALSE,
  as.Notch = FALSE, sort.by = NULL, sort.base = NULL,
  sort.ascending = FALSE, colorset = "black", symbol.color = "red",
  mean.symbol = 1, median.symbol = "|", outlier.symbol = 1,
  show.data = NULL, add.mean = TRUE, xlab = "Return",
  main = "Return Distribution Comparison", element.color = "darkgray", ...)
}
\description{
A wrapper to create box and whiskers plot with some defaults useful for
comparing distributions.
}
\details{
We have also provided controls for all the symbols and lines in the chart.
One default, set by \code{as.Tufte=TRUE}, will strip chartjunk and draw a
Boxplot per recommendations by Edward Tufte. Another default, set by \code{as.Notch=TRUE}, will draw a notch in each side of the boxes.  It can also be useful when comparing several series to sort them in the order of ascending or descending return or risk measurement by use of \code{sort.by} and  \code{sort.ascending=TRUE}. In addition, one can compare this with another user specified order, called base order, e.g., to see the relative change of the orders of the series between two measurements of interest.
}

