% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvo.constrained.r
\name{constraints}
\alias{constraints}
\title{Wrapper of constraints to constraint object}
\usage{
constraints(A, b, meq)
}
\description{
Wrapper of constraints to constraint object
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\seealso{
\code{\link{}}
}
\keyword{constraint}

