% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.sets.r
\name{cset.propcost}
\alias{cset.propcost}
\title{Constraint specifications}
\usage{
cset.propcost(returns, ptc, w.initial)
}
\description{
Constraint specifications
}
\seealso{
Other constraints: \code{\link{combine.cset}},
  \code{\link{cset.box}}, \code{\link{cset.fixed}},
  \code{\link{cset.groups}}, \code{\link{cset.lo}},
  \code{\link{cset.mu.target}}, \code{\link{cset.sum}},
  \code{\link{cset.turnover.hobbs}},
  \code{\link{cset.turnover}},
  \code{\link{propcost.modify}},
  \code{\link{turnover.hobbs.modify}},
  \code{\link{turnover.modify}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\concept{constraints}
\keyword{constraint}
