% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efront.constrained.r
\name{efrontPlot}
\alias{efrontPlot}
\title{Plot Efficient Frontier of Mean Variance Optimization}
\usage{
efrontPlot(returns, cset = NULL, mu.min = NULL, mu.max = NULL,
  rf = NULL, npoints = 10, wts.plot = T, wts.xlab = "VOL",
  printout = F, bar.ylim = c(0, 2), list.arg = NULL, is.fancy = TRUE)
}
\description{
Plot Efficient Frontier of Mean Variance Optimization
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\seealso{
\code{\link{efrontPlot}}
}
\keyword{efficient}
\keyword{frontier}

