% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvo.constrained.r
\name{gmv}
\alias{gmv}
\title{GMV Portfolio with Constraints}
\usage{
gmv(returns, cset = NULL, wts.only = T, digits = NULL)
}
\arguments{
\item{cset, }{if cset = NULL, then unconstrained gmv}

\item{wts.only, }{for back-test use default wts.only = T, for efficient frontier use wts.only = F}
}
\description{
GMV Portfolio with Constraints
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{constraint}
