% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathGmv.R
\name{mathGmv}
\alias{mathGmv}
\title{Global Minimum Variance (GMV) Portfolio}
\usage{
mathGmv(returns, digits = NULL)
}
\arguments{
\item{returns}{matrix of asset returns}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Compute the weights, mean return and volatility of a GMV portfolio.
}
\examples{
data(midcap.ts)
returns = midcap.ts[, 1:10]
mathGmv(returns)

}

