% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvo.constrained.r
\name{maxmu}
\alias{maxmu}
\title{Compute the Max Mean Return Portfolio with Constraints
This is primarily to compute the maximum mean return with constraints}
\usage{
maxmu(returns, cset, mu.only = TRUE, digits = NULL, verbose = FALSE)
}
\description{
Compute the Max Mean Return Portfolio with Constraints
This is primarily to compute the maximum mean return with constraints
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{constraint}
