% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvo.constrained.r
\name{minmu}
\alias{minmu}
\title{Compute the minimu of mean return using GMV}
\usage{
minmu(returns, cset = NULL)
}
\description{
Compute the minimu of mean return using GMV
}
\seealso{
\code{\link{}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{constraint}
