% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvo.constrained.r
\name{mvo}
\alias{mvo}
\title{MVO Portfolio with Constraints}
\usage{
mvo(returns, mu0, cset = NULL, wts.only = T, digits = NULL)
}
\description{
MVO Portfolio with Constraints
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\seealso{
\code{\link{}}
}
\keyword{constraint}

