% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraint.sets.r
\name{propcost.modify}
\alias{propcost.modify}
\title{Constraint specifications}
\usage{
propcost.modify(cset = cset.i)
}
\description{
Constraint specifications
}
\seealso{
Other constraints: \code{\link{combine.cset}},
  \code{\link{cset.box}}, \code{\link{cset.fixed}},
  \code{\link{cset.groups}}, \code{\link{cset.lo}},
  \code{\link{cset.mu.target}},
  \code{\link{cset.propcost}}, \code{\link{cset.sum}},
  \code{\link{cset.turnover.hobbs}},
  \code{\link{cset.turnover}},
  \code{\link{turnover.hobbs.modify}},
  \code{\link{turnover.modify}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\concept{constraints}
\keyword{constraint}
