% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.Performance.pool.R
\name{table.Performance.pool}
\alias{table.Performance.pool}
\title{Print metrics from R-forge PerformanceAnalytics that compatible with table.Performance}
\usage{
table.Performance.pool(...)
}
\description{
Print metrics from R-forge PerformanceAnalytics that compatible with table.Performance
}
\details{
use \code{table.Performance.pool} to check available metrics. recoded SharpeRatio
}
\examples{
table.Performance.pool()
}
\seealso{
\code{\link{table.Performance}}, \code{\link{table.Performance.pool.cran}}
}
\author{
Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{measure}
\keyword{metrics}
\keyword{performance}
\keyword{table}
