% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnoverOpt_doug.r
\name{TurnoverOpt_doug}
\alias{TurnoverOpt_doug}
\title{Specify turnover constraints using a version derived by Doug Martin}
\usage{
TurnoverOpt_doug(returns, mu.target = NULL, wts.initial, toc,
  long.only = TRUE, printDandA = FALSE)
}
\arguments{
\item{returns, }{stock returns in xts format.}

\item{mu.target, }{default is NULL, specify the target mean return}

\item{toc, }{the turnover cost}

\item{long.only, }{default is TRUE, specify if long only constraint is required}

\item{printDandA, }{default is FALSE, specify if print the D and A matrix}

\item{wts.inital, }{the initial weight}
}
\description{
Specify turnover constraints using a version derived by Doug Martin
}
\details{
use \code{table.Performance.pool} to check available metrics. recoded SharpeRatio
}
\author{
Doug Martin, Kirk Li  \email{kirkli@stat.washington.edu}
}
\keyword{constraints}
\keyword{turnover,}
